Separate out callback and conversions to internal modules
