+++ gi-gtk-declarative/gi-gtk-declarative.cabal
@@ -24,6 +24,8 @@ library
   exposed-modules:      GI.Gtk.Declarative
                       , GI.Gtk.Declarative.Attributes
                       , GI.Gtk.Declarative.Attributes.Internal
+                      , GI.Gtk.Declarative.Attributes.Internal.Callback
+                      , GI.Gtk.Declarative.Attributes.Internal.Conversions
                       , GI.Gtk.Declarative.Bin
                       , GI.Gtk.Declarative.Container
                       , GI.Gtk.Declarative.Container.Box
